+++ tests/Integration/ClientConfigurationProvider/BasicAuthTest.php
@@ -5,11 +5,11 @@ namespace Tests\Integration\ClientConfigurationProvider;
 use Sobak\Scrawler\Block\ClientConfigurationProvider\BasicAuthProvider;
 use Sobak\Scrawler\Block\Matcher\CssSelectorHtmlMatcher;
 use Sobak\Scrawler\Block\Matcher\CssSelectorListMatcher;
-use Sobak\Scrawler\Block\ResultWriter\InMemoryResultWriter;
 use Sobak\Scrawler\Configuration\ObjectConfiguration;
 use Sobak\Scrawler\Scrawler;
 use Tests\Integration\IntegrationTest;
 use Tests\Utils\BasicConfigurationProvider;
+use Tests\Utils\InMemoryResultWriter;
 use Tests\Utils\SimpleMatchEntity;
 
 /**